Rename to OctokitCLI
