+++ OctoKitCLI/Package.swift
@@ -3,7 +3,7 @@
 
 import PackageDescription
 
-let package = Package(name: "FixtureGenerator",
+let package = Package(name: "OctoKitCLI",
                       dependencies: [
                           .package(url: "https://github.com/apple/swift-argument-parser", from: "1.2.0"),
                           .package(url: "https://github.com/onevcat/Rainbow", from: "4.0.0"),
@@ -12,7 +12,7 @@ let package = Package(name: "FixtureGenerator",
                       targets: [
                           // Targets are the basic building blocks of a package. A target can define a module or a test suite.
                           // Targets can depend on other targets in this package, and on products in packages this package depends on.
-                          .executableTarget(name: "FixtureGenerator",
+                          .executableTarget(name: "OctoKitCLI",
                                             dependencies: [
                                                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
                                                 .product(name: "OctoKit", package: "octokit.swift"),
