+++ OctoKitCLI/Sources/OctoKitCLI/Modules/Issue.swift
@@ -11,8 +11,8 @@ import OctoKit
 import Rainbow
 
 @available(macOS 12.0, *)
-struct Gist: AsyncParsableCommand {
-    public static let configuration = CommandConfiguration(abstract: "Operate on Gists",
+struct Issue: AsyncParsableCommand {
+    public static let configuration = CommandConfiguration(abstract: "Operate on Issues",
                                                            subcommands: [
                                                                Get.self,
                                                                GetList.self
@@ -22,10 +22,16 @@ struct Gist: AsyncParsableCommand {
 }
 
 @available(macOS 12.0, *)
-extension Gist {
+extension Issue {
     struct Get: AsyncParsableCommand {
-        @Argument(help: "The id of the gist")
-        var id: String
+        @Argument(help: "The owner of the repository")
+        var owner: String
+
+        @Argument(help: "The name of the repository")
+        var repository: String
+
+        @Argument(help: "The number of the issue")
+        var number: Int
 
         @Argument(help: "The path to put the file in")
         var filePath: String?
@@ -37,19 +43,19 @@ extension Gist {
 
         mutating func run() async throws {
             let octokit = Octokit()
-            let session = FixtureURLSession()
-            _ = try await octokit.gist(session, id: id)
+            let session = JSONInterceptingURLSession()
+            _ = try await octokit.issue(session, owner: owner, repository: repository, number: number)
             session.verbosePrint(verbose: verbose)
             try session.printResponseToFileOrConsole(filePath: filePath)
         }
     }
-}
 
-@available(macOS 12.0, *)
-extension Gist {
     struct GetList: AsyncParsableCommand {
-        @Argument(help: "The owner of the gists")
+        @Argument(help: "The owner of the repository")
         var owner: String
+
+        @Argument(help: "The name of the repository")
+        var repository: String
 
         @Argument(help: "The path to put the file in")
         var filePath: String?
@@ -61,8 +67,8 @@ extension Gist {
 
         mutating func run() async throws {
             let octokit = Octokit()
-            let session = FixtureURLSession()
-            _ = try await octokit.gists(session, owner: owner)
+            let session = JSONInterceptingURLSession()
+            _ = try await octokit.issues(session, owner: owner, repository: repository)
             session.verbosePrint(verbose: verbose)
             try session.printResponseToFileOrConsole(filePath: filePath)
         }
