+++ OctoKitCLI/Sources/OctoKitCLI/Modules/PullRequest.swift
@@ -43,7 +43,7 @@ extension PullRequest {
 
         mutating func run() async throws {
             let octokit = Octokit()
-            let session = FixtureURLSession()
+            let session = JSONInterceptingURLSession()
             _ = try await octokit.pullRequest(session, owner: owner, repository: repository, number: number)
             session.verbosePrint(verbose: verbose)
             try session.printResponseToFileOrConsole(filePath: filePath)
@@ -67,7 +67,7 @@ extension PullRequest {
 
         mutating func run() async throws {
             let octokit = Octokit()
-            let session = FixtureURLSession()
+            let session = JSONInterceptingURLSession()
             _ = try await octokit.pullRequests(session, owner: owner, repository: repository)
             session.verbosePrint(verbose: verbose)
             try session.printResponseToFileOrConsole(filePath: filePath)
